switch: removing the need for sizing the container
In addition, the inital background divs look exactly like the two states of the
button so when a page slides in the switches have the right look for their
state.

Upon rendering, the fake buttons are all set to be transparent, and the real,
floating button takes over after being placed on top of the right fake button
for the current state.

At each state change, another floating button, itself transparent, is placed
above the fake buttons. Then, it's coordinates are retrieved and used to animate
the real button into place.