[theming] lessify the jqm theme CSS file
After further consideration, it seemed like a good idea to
lessify our theme file, as it helps us to understand the make-up
of a jqm theme better. It also makes it easier to tweak colours,
as they are all in one place (in variables at the top of each
swatch).

So far, I've only done this for the 'a' and 'c' swatches, as these
are the default ones and so most frequently applied.