[theming] Prevent clipping of slider handle edges
The changes to make the background color correct for button
list items made the edges of the switch slider handle clip (some
CSS interaction which made it difficult to fix for just
the slider without breaking the buttons again).

This commit fixes the clipping by sliding the handle slightly
further over to the left or right. This is not noticeable visually
(even though it means the slider is slightly off-centre) but
avoids the clipping.