Tweaking things to make them look more like the spec
There appears to be a top and bottom border around screen-wide
widgets such as hsvpicker, datetimepicker, colorpalette, and
colortitle. This means that if stacked, and if both top and bottom
border is specified, there will be a double border between widgets.

To minimize styling, each widget is created with only a bottom
border. Then, to "put a lid" on a stack of widgets, use a label,
which is created with a bottom border identical to each widget's.