[optionheader] Add CSS for segmented controls
Still not sure about this approach, as it basically involves
undoing all the CSS applied by jqm. It also doesn't work well
for segmented controls with icons positioned at the top of the
buttons (the buttons push out the bottom margin under the
control); and the sizing of the buttons doesn't match the
spec (it's left to jqm).