[theming] Remove bottom border from header
Previously I set a 5px border on the bottom of the header
to provide the thick orange line when option header is collapsed.

However, this makes positioning the option header with respect
to the header more difficult: it has to account for the border
at the bottom of the header, which may differ between themes. This
is made more difficult where the option header requires an
upward pointing triangle (the indicator showing that content
is hidden inside a collapsible area).

Instead, the theme is now responsible for positioning the option
header if there is a default border on the bottom of the header. In
cases where there is no border (which is the default for this theme),
the option header will appear correctly, and display the upward
pointing triangle correctly.