[optionheader] Use swatch 'b' and derive arrow color from background
Keeping the arrow color in keeping with the background color
from the bar is difficult: there isn't a swatch which just applies
borders, so using ui-bar-c (or 'a' or 'b') was filling in a block
behind the arrow.

To work around this, derive the border color for the arrow from
the background color of the option header.

Also change the default swatch for the option header to 'b'.