[slider] Remove popup when slider stops drag
Attach to the vmouse events on the document (following the
code in the jqm slider) to determine when the slider has
been released.

While the slider is being dragged, the popup is visible; once
released, the popup disappears.

NB the popup makes space for itself in the page when
it's made visible (due to the addition of the ui-slider-handle
class to the popup). This could be removed, in which case
the popup will be transparent, but will overlay other elements;
in cases where there isn't room above the slider, the popup
will appear below it.