shortcutscroll: mouse works on the emulator
On the emulator, the element whereupon the initial mousedown happend
seems to capture the mouse thus receiving all subsequent mouse events.
Of course, no mouse events are issued when the pointer floats above
a certain div - it wouldn't make sense on a touch device.

Thus, the mouse handling had to be changed to always convert to
shortcuts list coordinates and then hit-test individual list items
to identify which one the mouse is currently in.