[personpicker] Init optionheader after DOM element appended
Because the optionheader is being added to a "floating" DOM fragment
and not directly to a page, it can't bind to the pageshow for the
containing page.

By moving the call to optionheader(), we can be sure that the DOM
is attached to the page, so the optionheader can find it.