[scrollview] Stop event blocking for input and link elements
The scrollview blocks event propagation for mouse events
(not touch events). This stops input and link elements inside
a scrollview from working.

Amend the condition on blocking, so mouse event
propagation/default behaviour still applies if the event occurs
on a link or input element.

NB this needs to go upstream, as it's an edit on the
experimental scrollview code.