[optionheader] Only measure optionheader height if it's not set
The optionheader was being measured whenever the page was displayed.
This meant that if the page had already been displayed, and
the optionheader collapsed, it would get measured again in its
collapsed state. Which would then mean it couldn't be expanded.

To fix, don't measure the optionheader if a value has already
been determined for its height (this will always be the expanded
height the first time the optionheader is displayed, as it
always starts out expanded; if set to collapse on init, it is
measured first before it is collapsed).