[theming] Copy widget CSS into each theme built
As the widget CSS shouldn't be theme specific, but the images for
a widget may need to be, copy the CSS for all widgets into each
of the generated themes.

This introduces redundancy, but it means that widget CSS can
correctly reference images specific to a theme (as the widget
CSS file will be co-located with the theme CSS file, and have
the same path to a theme's images directory).