[personpicker] Move stubs and images to demo
The stub and images for the avatar properly belong in demo.

If we decide to provide a default avatar for a person (e.g.
they don't have one), we should put the avatar PNG into the
theme directories. However, at the moment, we're just using
it as the image for every stub contact, which is a different
thing; so the image should go into the demo.

This also necessitated changing personpicker slightly, so
the stub address book can be applied to it after _create time.