[hbox] Retrieve children correctly after scrollview applied
The children of the element where the hbox applies are wrapped
inside a scrollview div after the scrollview is applied. So
the first time the widget is refreshed, everything is OK;
however, on subsequent refreshes, the scrollview view was the
only child, so all the calculations went wrong.

Make sure that the items to lay out are the correct, original
set of children. This has to be calculated on each refresh,
in case the set of children has changed.