Set jqmData as well as adding data-role attribute
The shortcutscroll looks for data-role attributes on
listview element to determine where list dividers are;
but these aren't set by jqmData.

However, tests can use data-role attributes, as they are under
a different namespace; so they need jqmData to be set on each
list divider.

So this commit adds a data-role attribute so that the listview
can generate dividers correctly, which can then be picked up
by shortcutscroll.