[gallery] Add some more list items to the popupwindow demo
Added some more li items to the popupwindow list demo, but
the styling is unsatisfactory (the second item's background
appears to be truncated).

It looks like adding some padding to the ui-li class declaration
in the CSS might fix this (though it needs testing in all
listview contexts).

Might be worth waiting until we are using jQuery Mobile CSS
(see #191) and see whether fixing it there breaks anything else. Also
need to investigate why the styling is wrong for listviews when
they're inside a popupwindow in the first place.

Fixes #184