[swipelist] Make styling more generic
Copy styling from the li item containing the cover onto the cover.
This means the cover looks like the list item it's covering (by
default).

Move all the styling specific to the particular swipelist covers
in the demo into the demo CSS.

Wrap the HTML of the cover in a span with no styling, which can
then be styled as desired.