[listviewcontrols] Change how test executes
Because QUnit stops the test at the point where an assertion
fails, the destroy() wasn't being called if a the equal()
assertion was failing. This then made subsequent tests fail.

To fix, store the result of the hasClass() test in a variable,
then destroy the widget before doing the assertion.