[shortcutscrol] Set height for scrollview clip based on shortcuts list
Ensure that the height of the scrollview clip is sufficient so
that all of the links in the shortcut list are visible.

This is particularly important for personpicker_page, where we
need the shortcutscroll to dictate the size of the personpicker
(which in turn sets the size of the personpicker_page).