[dayselector] Don't uppercase label text
Some languages (e.g. French) don't use title case for names of
days of the week. So the text on the dayselector labels shouldn't
be set to uppercase.

This can be set to uppercase by passing a custom "days" option
to the constructor where days of the week are in title case.