[swipelist] Properly unwrap children of a cover
The old code for unwrapping children of a cover (to remove
markup applied by the widget) didn't properly account for
covers with text-only (no HTML elements) inside them. Those
covers were getting a span wrapped round the text, which
wasn't being removed by destroy().

This commit fixes it so that if the content of a cover is
text, and the text is inside the wrapper span, append the text
back onto the cover and remove the span.